requirements.txt
aiodns==3.0.0
aiofiles==22.1.0
aiohttp==3.8.3
aioschedule==0.5.2
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
boto3==1.26.100
botocore==1.29.100
Brotli==1.0.9
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
discord.py==2.0.1
fonttools==4.38.0
frozenlist==1.3.1
google-api-core==2.11.0
google-api-python-client==2.65.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-cloud-vision==3.3.1
googleapis-common-protos==1.56.4
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
jmespath==1.0.1
kiwisolver==1.4.4
matplotlib==3.6.2
multidict==6.0.2
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
proto-plus==1.22.2
protobuf==4.21.9
psycopg2==2.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==4.2.2
pycparser==2.21
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
requests-ip-rotator==1.0.14
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
schedule==1.1.0
seaborn==0.12.1
six==1.16.0
soupsieve==2.3.2.post1
tabulate==0.9.0
uritemplate==4.1.1
urllib3==1.26.12
yarl==1.8.1
